Adding authentication to REST side
Allowing authentication on REST APIs. When the REST API layer calls into
the meteor methods the userId and username are not available. Not sure
how to set these so instead passing in an object from REST to Methods
that includes the userId and username. The methods can now check this
object for the values and set them properly if they are received. All
Meteor methods should be updated to use these fields now.